16/1/2025

+++ VehicleManagement/Controllers/UsersShiftController.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System.Data;
 using VehicleManagement.Classes;
@@ -13,6 +12,7 @@ namespace VehicleManagement.Controllers
     public class UsersShiftController : ControllerBase
     {
         private readonly ConnectionClass _connection;
+        private apiResponse Resp = new apiResponse();
         LkDataConnection.DataAccess _dc = new LkDataConnection.DataAccess();
         LkDataConnection.SqlQueryResult _query = new LkDataConnection.SqlQueryResult();
         public UsersShiftController(ConnectionClass connection)
@@ -27,34 +27,49 @@ namespace VehicleManagement.Controllers
         [Route("GetAllShifts")]
         public IActionResult GetAllShifts()
         {
-            string query = $"select * from User_Shift_Mst  ORDER BY Shift_Name ASC";
-            Console.WriteLine(query);
-            var connection = new LkDataConnection.Connection();
-
-            var result = connection.bindmethod(query);
-
-
-            DataTable Table = result._DataTable;
-
-            var ShiftList = new List<UsershiftModel>();
-
-            foreach (DataRow row in Table.Rows)
+            try
             {
-                ShiftList.Add(new UsershiftModel
+                string query = $"select * from User_Shift_Mst  ORDER BY Shift_Name ASC";
+                Console.WriteLine(query);
+                var connection = new LkDataConnection.Connection();
+
+                var result = connection.bindmethod(query);
+
+
+                DataTable Table = result._DataTable;
+
+                var ShiftList = new List<UsershiftModel>();
+
+                foreach (DataRow row in Table.Rows)
                 {
-                    Shift_Id = Convert.ToInt32(row["Shift_Id"]),
-                    Shift_Name = row["Shift_Name"].ToString(),
-                    Start_Time = TimeSpan.Parse(row["Start_Time"].ToString()), 
-                    End_Time = TimeSpan.Parse(row["End_Time"].ToString()),
-                    Grace_Time = float.Parse(row["Grace_Time"].ToString()),
-                    Created_At = Convert.ToDateTime(row["Created_At"]),
-                    Shift_Status = Convert.ToBoolean(row["Shift_Status"])
+                    ShiftList.Add(new UsershiftModel
+                    {
+                        Shift_Id = Convert.ToInt32(row["Shift_Id"]),
+                        Shift_Name = row["Shift_Name"].ToString(),
+                        End_Time = TimeSpan.Parse(row["End_Time"].ToString()).ToString("hh\\:mm"),
+                        Start_Time = TimeSpan.Parse(row["Start_Time"].ToString()).ToString("hh\\:mm"),
+                        Grace_Time = float.Parse(row["Grace_Time"].ToString()),
+                        Created_At = Convert.ToDateTime(row["Created_At"]),
+                        Shift_Status = Convert.ToBoolean(row["Shift_Status"])
 
-                });
+                    });
+                }
+
+                Resp.StatusCode = StatusCodes.Status200OK;
+                Resp.Message = $"Shift List fetched successfully ";
+                Resp.ApiResponse = ShiftList;
+                Resp.IsSuccess = true;
+
+                return Ok(Resp);
             }
+            catch (Exception ex)
+            {
+                Resp.StatusCode = StatusCodes.Status500InternalServerError;
+                Resp.Message = ex.Message;
 
 
-            return Ok(ShiftList);
+                return StatusCode(StatusCodes.Status500InternalServerError, Resp);
+            }
         }
 
         [HttpPost]
@@ -74,23 +89,41 @@ namespace VehicleManagement.Controllers
                 bool isDuplicate = duplicacyChecker.CheckDuplicate(duplicacyParameter);
                 if (isDuplicate)
                 {
-                    return StatusCode(StatusCodes.Status208AlreadyReported, new { message = "Shift already exists.", DUP = true });
+                    Resp.StatusCode = StatusCodes.Status208AlreadyReported;
+                    Resp.Message = $"Shift already exists.";
+                    Resp.Dup = true;
+
+                    return StatusCode(StatusCodes.Status208AlreadyReported, Resp);
                 }
                 if (String.IsNullOrEmpty(Usershift.Shift_Name))
                 {
-                    return StatusCode(StatusCodes.Status200OK, new { message = "Shift Can't be Blank Or Null", DUP = false });
-                }
-                Usershift.Start_Time = new TimeSpan(Usershift.Start_Time.Value.Hours, Usershift.End_Time.Value.Minutes, 0);
+                    Resp.StatusCode = StatusCodes.Status404NotFound;
+                    Resp.Message = $"Shift Can't be Blank Or Null";
 
-                Usershift.End_Time = new TimeSpan(Usershift.End_Time.Value.Hours, Usershift.End_Time.Value.Minutes, 0);
+                    return StatusCode(StatusCodes.Status404NotFound, Resp);
+                }
+
+                //Usershift.Start_Time = new TimeSpan(Usershift.Start_Time.Value.Hours, Usershift.End_Time.Value.Minutes, 0);
+
+                //Usershift.End_Time = new TimeSpan(Usershift.End_Time.Value.Hours, Usershift.End_Time.Value.Minutes, 0);
 
                 _query = _dc.InsertOrUpdateEntity(Usershift, "User_Shift_Mst", -1);
-                return StatusCode(StatusCodes.Status200OK, new { message = "Shift Added Successfully", DUP = false });
+                Resp.StatusCode = StatusCodes.Status200OK;
+                Resp.Message = "Shift Added Successfully";
+                Resp.IsSuccess = true;
+
+
+                return StatusCode(StatusCodes.Status200OK, Resp);
+
 
             }
             catch (Exception ex)
             {
-                return StatusCode(StatusCodes.Status500InternalServerError, $"Error: {ex.Message}");
+                Resp.StatusCode = StatusCodes.Status500InternalServerError;
+                Resp.Message = ex.Message;
+
+
+                return StatusCode(StatusCodes.Status500InternalServerError, Resp);
             }
         }
         [HttpPut]
@@ -107,7 +140,7 @@ namespace VehicleManagement.Controllers
 
                 if (result == 0)
                 {
-                    return StatusCode(StatusCodes.Status404NotFound, new { message = "Role ID does not exist.", DUP = false });
+                    return StatusCode(StatusCodes.Status404NotFound, new { message = "Shift ID does not exist.", DUP = false });
                 }
 
                 var duplicacyChecker = new CheckDuplicacy(_connection);
@@ -125,19 +158,35 @@ namespace VehicleManagement.Controllers
 
                 if (isDuplicate)
                 {
-                    return StatusCode(StatusCodes.Status208AlreadyReported, new { message = "Shift Name already exists.", DUP = true });
+                    Resp.StatusCode = StatusCodes.Status208AlreadyReported;
+                    Resp.Message = $"Shift already exists.";
+                    Resp.Dup = true;
 
+                    return StatusCode(StatusCodes.Status208AlreadyReported, Resp);
                 }
-                usershift.Start_Time = new TimeSpan(usershift.Start_Time.Value.Hours, usershift.End_Time.Value.Minutes, 0);
 
-                usershift.End_Time = new TimeSpan(usershift.End_Time.Value.Hours, usershift.End_Time.Value.Minutes, 0);
+
+
+
+                //usershift.Start_Time = new TimeSpan(usershift.Start_Time.Value.Hours, usershift.End_Time.Value.Minutes, 0);
+
+                //usershift.End_Time = new TimeSpan(usershift.End_Time.Value.Hours, usershift.End_Time.Value.Minutes, 0);
                 _query = _dc.InsertOrUpdateEntity(usershift, "User_Shift_Mst", Shift_Id, "Shift_Id");
-                return StatusCode(StatusCodes.Status200OK, new { message = "Shift Name Updated Successfully", DUP = false });
+                Resp.StatusCode = StatusCodes.Status200OK;
+                Resp.Message = "Shift Name Updated Successfully";
+                Resp.IsSuccess = true;
+
+
+                return StatusCode(StatusCodes.Status200OK, Resp);
 
             }
             catch (Exception ex)
             {
-                return StatusCode(StatusCodes.Status500InternalServerError, $"Error{ex.Message}");
+                Resp.StatusCode = StatusCodes.Status500InternalServerError;
+                Resp.Message = ex.Message;
+
+
+                return StatusCode(StatusCodes.Status500InternalServerError, Resp);
             }
         }
 
@@ -159,18 +208,39 @@ namespace VehicleManagement.Controllers
                 {
                     return StatusCode(StatusCodes.Status404NotFound, new { message = "Shift ID does not exist.", DUP = false });
                 }
+                string checkQuery = $"SELECT COUNT(*) AS recordCount FROM User_mst WHERE Shift_Id = {id}";
 
-               
+
+                int shift_IdInUser = Convert.ToInt32(_connection.ExecuteScalar(checkQuery));
+                if (shift_IdInUser > 0)
+                {
+                    Resp.StatusCode = StatusCodes.Status404NotFound;
+                    Resp.Message = $"Can't delete Exists in another table ";
+
+                    return StatusCode(StatusCodes.Status404NotFound, Resp);
+                  
+
+
+                }
+
                 string deleteRoleQuery = $"Delete from User_Shift_Mst where Shift_Id ='{id}'";
 
                 LkDataConnection.Connection.ExecuteNonQuery(deleteRoleQuery);
-                return StatusCode(StatusCodes.Status200OK, new { message = "Shift  Deleted successfully" });
+                Resp.StatusCode = StatusCodes.Status200OK;
+                Resp.Message = "Shift  Deleted successfully";
+                Resp.IsSuccess = true;
 
+
+                return StatusCode(StatusCodes.Status200OK, Resp);
 
             }
             catch (Exception ex)
             {
-                return StatusCode(StatusCodes.Status500InternalServerError, $"Error{ex.Message}");
+                Resp.StatusCode = StatusCodes.Status500InternalServerError;
+                Resp.Message = ex.Message;
+
+
+                return StatusCode(StatusCodes.Status500InternalServerError, Resp);
             }
         }
